requirements.txt
gradio
sentence-transformers
faiss-cpu
numpy
pandas